The Ares Poster Now Correctly Calls The Planet Mars (#3193)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

Does what it says on the title.

![image](https://github.com/shiptest-ss13/Shiptest/assets/77556824/dc713a4f-685f-44d5-a73b-3a4e3cc24da1)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

It was a weird misconception that for some reason SolCon decided to
rename all the planets into their greek counterparts. This is very much
not the case, especially since if it *was*, Terra would've been called
Gaia and the moon would've been called Selene instead of Luna.

This does however mean that Uranus is now called Caelus, though this
will be depicted in a poster in another time.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl: PositiveEntropy
fix: Mars and Venus are now referred to by their correct names!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->