SolCon/SolGov Hardsuits (And Its Subtypes) Can Now Stow Away Guns! (#3187)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

Does what it says on the tin. You can now holster guns, ammo and such on
the armor slot while equipping the hardsuit.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

I almost got shot because people got terrified at the fact I had a gun
on my hand, when I couldn't holster due to this oversight. Amazing.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl: PositiveEntropy
fix: You can now store guns, ammunition, and melee weapons on the armor
slot when equipping a SolCon/SolGov hardsuit, and its subtypes!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->