Gloves now show up in the loadout preview. (#3215)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
If you select gloves in the loadout, they now show up on your character.
Code-wise, this seperates gloves into a new category.

Oh and also, adds brown gloves to the loadout.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
It's nice to see what your character is meant to be wearing.
Brown gloves were added because they can totally fit with some drip,
while not being bright and flashy, unlike some other gloves.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl:
fix: Gloves now show up in the loadout preview.
add: Brown gloves were added to the loadout.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->